+++ src/hipp/kh9pc/fiducial_patterns.py
