+++ src/hipp/kh9pc/kh9_image_spec.py
@@ -5,7 +5,7 @@ import re
 import rasterio
 
 
-from hipp.kh9pc.fiducials import (
+from hipp.kh9pc.fiducial_patterns import (
     FiducialPattern,
     RegularDense,
     RegularMid,
