+++ src/hipp/kh9pc/pipeline.py
@@ -7,8 +7,8 @@ import matplotlib.pyplot as plt
 
 from hipp.image import generate_quickview
 from hipp.kh9pc.mosaic import image_mosaic
-from hipp.kh9pc.qc import get_figures
-from hipp.kh9pc.restitution.mixed import MixedStrategy
+from hipp.kh9pc.quality_control import get_figures
+from hipp.kh9pc.restitution.mixed_strategy import MixedStrategy
 from hipp.tools import extract_archive
 
 logger = logging.getLogger(__name__)
