--- src/hipp/kh9pc/qc/collimation.py
@@ -1,45 +1,0 @@
-import matplotlib.pyplot as plt
-import numpy as np
-from matplotlib.figure import Figure
-
-from hipp.kh9pc.restitution.collimation import CollimationStrategy
-
-
-def plot_collimation_edges(detector: CollimationStrategy) -> Figure:
-    """Subimage thumbnails with RANSAC inliers/outliers and polynomial model for top and bottom collimation lines."""
-    fig, axes = plt.subplots(1, 2, figsize=(8, 4), constrained_layout=True)
-
-    for ax, side, result in zip(axes, ["top", "bottom"], [detector.top_, detector.bottom_]):
-        ax.imshow(result.sub_image.band, cmap="gray", aspect="auto")
-
-        inliers = result.model.inlier_mask_
-        peaks = result.peaks_local
-        ax.scatter(peaks[~inliers, 0], peaks[~inliers, 1], s=12, c="red", label="outliers")
-        ax.scatter(peaks[inliers, 0], peaks[inliers, 1], s=12, c="green", label="inliers")
-
-        y_global_pred = result.model.predict(result.peaks_global[:, 0].reshape(-1, 1))
-        global_pred = np.column_stack([result.peaks_global[:, 0], y_global_pred])
-        local_pred = result.sub_image.to_local(global_pred)
-        ax.plot(local_pred[:, 0], local_pred[:, 1], color="blue", linewidth=1, label="model")
-
-        ax.set_title(f"{side} collimation line")
-        ax.legend(loc="best", fontsize=8)
-        ax.axis("off")
-
-    return fig
-
-
-def plot_collimation_distortions(detector: CollimationStrategy) -> Figure:
-    """Residual distortion curves for top and bottom collimation line fits."""
-    fig, ax = plt.subplots(figsize=(6, 4), constrained_layout=True)
-
-    for side, result in zip(["top", "bottom"], [detector.top_, detector.bottom_]):
-        ax.plot(result.distortion[:, 0], result.distortion[:, 1], label=side)
-
-    ax.invert_yaxis()
-    ax.legend()
-    ax.set_title("global distortion (top & bottom)")
-    ax.set_xlabel("column (px)")
-    ax.set_ylabel("distortion (px)")
-
-    return fig
