--- src/hipp/kh9pc/qc/poly.py
@@ -1,46 +1,0 @@
-import matplotlib.pyplot as plt
-import numpy as np
-from matplotlib.figure import Figure
-
-from hipp.kh9pc.restitution.poly import PolyStrategy
-
-
-def plot_poly_edges(detector: PolyStrategy) -> Figure:
-    """Subimage thumbnails with RANSAC inliers/outliers and polynomial model for top and bottom edges."""
-    fig, axes = plt.subplots(1, 2, figsize=(8, 4), constrained_layout=True)
-
-    for ax, side, result in zip(axes, ["top", "bottom"], [detector.top_, detector.bottom_]):
-        ax.imshow(result.sub_image.band, cmap="gray", aspect="auto")
-
-        inlier_mask = result.model.inlier_mask_
-        pts = result.ruptures_local
-        ax.scatter(pts[~inlier_mask, 0], pts[~inlier_mask, 1], s=12, c="red", label="outliers")
-        ax.scatter(pts[inlier_mask, 0], pts[inlier_mask, 1], s=12, c="green", label="inliers")
-
-        x_global = result.ruptures_global[:, 0].astype(float)
-        y_global_pred = result.model.predict(x_global.reshape(-1, 1))
-        global_pred = np.column_stack([x_global, y_global_pred.ravel()])
-        local_pred = result.sub_image.to_local(global_pred)
-        ax.plot(local_pred[:, 0], local_pred[:, 1], color="blue", linewidth=1, label="model")
-
-        ax.set_title(f"{side} edge")
-        ax.legend(loc="best", fontsize=8)
-        ax.axis("off")
-
-    return fig
-
-
-def plot_poly_distortions(detector: PolyStrategy) -> Figure:
-    """Residual distortion curves (deviation from mean) for top and bottom polynomial fits."""
-    fig, ax = plt.subplots(figsize=(6, 4), constrained_layout=True)
-
-    ax.plot(detector.top_.distortion[:, 0], detector.top_.distortion[:, 1], label="top")
-    ax.plot(detector.bottom_.distortion[:, 0], detector.bottom_.distortion[:, 1], label="bottom")
-    ax.axhline(0, color="gray", linewidth=0.8, linestyle="--")
-    ax.invert_yaxis()
-    ax.legend()
-    ax.set_title("global distortion (top & bottom)")
-    ax.set_xlabel("column (px)")
-    ax.set_ylabel("distortion (px)")
-
-    return fig
