--- src/hipp/kh9pc/qc/vertical.py
@@ -1,49 +1,0 @@
-import matplotlib.pyplot as plt
-import rasterio
-from matplotlib.figure import Figure
-from rasterio.warp import Resampling
-from rasterio.windows import Window
-
-from hipp.kh9pc.restitution.vertical import VerticalDetector
-
-
-def plot_vertical_ruptures(detector: VerticalDetector) -> Figure:
-    """Band profiles with detected rupture positions for left and right edges."""
-    fig, axes = plt.subplots(1, 2, figsize=(8, 4), constrained_layout=True)
-
-    for ax, side, result in zip(axes, ["left", "right"], [detector.left_, detector.right_]):
-        profile = result.sub_image.band.flatten()
-        ax.plot(profile, color="gray")
-        ax.axvline(x=result.rupture_local, color="red", label=f"rupture (local={result.rupture_local})")
-        ax.set_title(f"{side} band profile (global col={result.position})\nGradient:{result.gradient_pct:.2%}")
-        ax.set_xlabel("local column index")
-        ax.set_ylabel("intensity")
-        ax.legend()
-
-    return fig
-
-
-def plot_vertical_edges(
-    detector: VerticalDetector,
-    margin_fraction: float = 0.03,
-    plot_res: float = 0.05,
-) -> Figure:
-    """Thumbnails around the left and right edge positions."""
-    fig, axes = plt.subplots(1, 2, figsize=(8, 4), constrained_layout=True)
-
-    with rasterio.open(detector.raster_filepath_) as src:
-        margin = int(src.width * margin_fraction)
-
-        for ax, side, edge_col in zip(axes, ["left", "right"], detector.edges_):
-            col_off = max(0, edge_col - margin)
-            col_end = min(src.width, edge_col + margin)
-            window = Window(col_off, 0, col_end - col_off, src.height)
-            out_shape = (1, int(src.height * plot_res), int(window.width * plot_res))
-            band = src.read(1, window=window, out_shape=out_shape, resampling=Resampling.average)
-
-            ax.imshow(band, cmap="gray", aspect="auto")
-            ax.axvline(x=(edge_col - col_off) * plot_res, color="red")
-            ax.set_title(f"{side} edge (col={edge_col})")
-            ax.axis("off")
-
-    return fig
