+++ src/hipp/kh9pc/restitution/collimation_strategy.py
@@ -12,9 +12,9 @@ from skimage.transform import ThinPlateSplineTransform
 from sklearn.linear_model import RANSACRegressor
 
 from hipp.image import SubImage, remap_tif_blockwise
-from hipp.kh9pc.fitting import fit_ransac_poly
+from hipp.kh9pc.restitution.base import fit_ransac_poly
 from hipp.kh9pc.restitution.base import DEFAULT_OUTPUT_HEIGHT, RestitutionStrategy, Transformation
-from hipp.kh9pc.restitution.poly import PolyStrategy
+from hipp.kh9pc.restitution.poly_strategy import PolyStrategy
 
 
 @dataclass
