+++ src/hipp/kh9pc/restitution/poly_strategy.py
@@ -10,9 +10,9 @@ from skimage.transform import ThinPlateSplineTransform
 from sklearn.linear_model import RANSACRegressor
 
 from hipp.image import SubImage, remap_tif_blockwise
-from hipp.kh9pc.fitting import detect_ruptures, fit_ransac_poly
+from hipp.kh9pc.restitution.base import detect_ruptures, fit_ransac_poly
 from hipp.kh9pc.restitution.base import DEFAULT_OUTPUT_HEIGHT, RestitutionStrategy, Transformation
-from hipp.kh9pc.restitution.vertical import VerticalDetector
+from hipp.kh9pc.restitution.vertical_detector import VerticalDetector
 
 
 @dataclass
